Nie ma wpisywania adresów z palca!

+++ Silownia/Silownia/Controllers/KlientController.cs
@@ -23,65 +23,76 @@ namespace Silownia.Controllers
  
         public ActionResult Index(string Miasto,string imieNazwisko, bool czyUmowa =false ,int page=1 ,int pageSize = 10 , AkcjaEnum akcja = AkcjaEnum.Brak , String info = null)
         {
-
-            var Miasta = db.Klienci.Where(u => (u.OsobaID != null) && (u.Adres != null)).DistinctBy(a => new { a.Adres.Miasto }).Select(x => x.Adres);
-
-
-            ViewBag.Miasto = new SelectList(Miasta, "Miasto", "Miasto");
-            //ViewBag.srchImieNazwisko = imieNazwisko;   Póki co niech będzie zakomentowane 
-            //ViewBag.czyUmowa = czyUmowa;
-   
-            var osoby = from Osoby in db.Klienci select Osoby;
-
-
-            if (!String.IsNullOrEmpty(imieNazwisko))
-                foreach (string wyraz in imieNazwisko.Split(' '))
-                    osoby = osoby.Search(wyraz, i => i.Imie, i => i.Nazwisko);
-            osoby = osoby.Search(Miasto, m => m.Adres.Miasto);
-            if(czyUmowa)
-            osoby = osoby.Where(u => u.Umowy.Count > 0);
-          
-            var final = osoby.OrderBy(p => p.Imie);
-            var ileWynikow = osoby.Count();
-            if ((ileWynikow / page) <= 1)
+            if (Session["User"] != null)
             {
-                page = 1;
-            }
-            var kk = ileWynikow / page;
+                var Miasta = db.Klienci.Where(u => (u.OsobaID != null) && (u.Adres != null)).DistinctBy(a => new { a.Adres.Miasto }).Select(x => x.Adres);
+
+
+                ViewBag.Miasto = new SelectList(Miasta, "Miasto", "Miasto");
+                //ViewBag.srchImieNazwisko = imieNazwisko;   Póki co niech będzie zakomentowane 
+                //ViewBag.czyUmowa = czyUmowa;
+
+                var osoby = from Osoby in db.Klienci select Osoby;
+
+
+                if (!String.IsNullOrEmpty(imieNazwisko))
+                    foreach (string wyraz in imieNazwisko.Split(' '))
+                        osoby = osoby.Search(wyraz, i => i.Imie, i => i.Nazwisko);
+                osoby = osoby.Search(Miasto, m => m.Adres.Miasto);
+                if (czyUmowa)
+                    osoby = osoby.Where(u => u.Umowy.Count > 0);
+
+                var final = osoby.OrderBy(p => p.Imie);
+                var ileWynikow = osoby.Count();
+                if ((ileWynikow / page) <= 1)
+                {
+                    page = 1;
+                }
+                var kk = ileWynikow / page;
 
                 PagedList<Klient> model = new PagedList<Klient>(final, page, pageSize);
 
 
-            if(akcja != AkcjaEnum.Brak)
-            {
-                ViewBag.info = info;
-                ViewBag.Akcja = akcja;
+                if (akcja != AkcjaEnum.Brak)
+                {
+                    ViewBag.info = info;
+                    ViewBag.Akcja = akcja;
+                }
+
+                return View(model);
             }
- 
-            return View(model);
+            return HttpNotFound();
 
         }
        
         // GET: /Klient/Details/5
         public ActionResult Details(long? id)
         {
-            if (id == null)
+            if (Session["User"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Klient klient = db.Klienci.Find(id);
+                if (klient == null)
+                {
+                    return HttpNotFound();
+                }
+                var z = klient;
+                return View(z);
             }
-            Klient klient = db.Klienci.Find(id);
-            if (klient == null)
-            {
-                return HttpNotFound();
-            }
-            var z = klient;
-            return View(z);
+            return HttpNotFound();
         }
 
         // GET: /Klient/Create
         public ActionResult Create()
         {
-            return View();
+            if (Session["User"] != null)
+            {
+                return View();
+            }
+            return HttpNotFound();
         }
 
         // POST: /Klient/Create
@@ -91,39 +102,45 @@ namespace Silownia.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "OsobaID,Imie,Nazwisko,DataUrodzenia,Mail,NrTelefonu,Adres")] Klient klient)
         {
-           
-
-            if (ModelState.IsValid)
+            if (Session["User"] != null)
             {
-                db.Klienci.Add(klient);
-                db.SaveChanges();
+                if (ModelState.IsValid)
+                {
+                    db.Klienci.Add(klient);
+                    db.SaveChanges();
 
-                Uzytkownik uzytkownik = new Uzytkownik();
-                uzytkownik.IDOsoby = klient.OsobaID;
-                uzytkownik.Login = klient.Mail;
-                uzytkownik.Haslo = klient.Imie + klient.Nazwisko;
-                uzytkownik.Rola = "Klient";
-                db.Uzytkownicy.Add(uzytkownik);
-                db.SaveChanges();
-                return RedirectToAction("Index", new { akcja = AkcjaEnum.DodanoKlienta , info = klient.imieNazwisko });
+                    Uzytkownik uzytkownik = new Uzytkownik();
+                    uzytkownik.IDOsoby = klient.OsobaID;
+                    uzytkownik.Login = klient.Mail;
+                    uzytkownik.Haslo = klient.Imie + klient.Nazwisko;
+                    uzytkownik.Rola = "Klient";
+                    db.Uzytkownicy.Add(uzytkownik);
+                    db.SaveChanges();
+                    return RedirectToAction("Index", new { akcja = AkcjaEnum.DodanoKlienta, info = klient.imieNazwisko });
+                }
+
+                return View(klient);
             }
-            
-            return View(klient);
+            return HttpNotFound();
         }
 
         // GET: /Klient/Edit/5
         public ActionResult Edit(long? id)
         {
-            if (id == null)
+            if (Session["User"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Klient klient = db.Klienci.Find(id);
+                if (klient == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(klient);
             }
-            Klient klient = db.Klienci.Find(id);
-            if (klient == null)
-            {
-                return HttpNotFound();
-            }
-            return View(klient);
+            return HttpNotFound();
         }
 
         // POST: /Klient/Edit/5
@@ -133,28 +150,36 @@ namespace Silownia.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "OsobaID,Imie,Nazwisko,DataUrodzenia,Mail,NrTelefonu")] Klient klient)
         {
-            if (ModelState.IsValid)
+            if (Session["User"] != null)
             {
-                db.Entry(klient).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                if (ModelState.IsValid)
+                {
+                    db.Entry(klient).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                return View(klient);
             }
-            return View(klient);
+            return HttpNotFound();
         }
 
         // GET: /Klient/Delete/5
         public ActionResult Delete(long? id)
         {
-            if (id == null)
+            if (Session["User"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Klient klient = db.Klienci.Find(id);
+                if (klient == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(klient);
             }
-            Klient klient = db.Klienci.Find(id);
-            if (klient == null)
-            {
-                return HttpNotFound();
-            }
-            return View(klient);
+            return HttpNotFound();
         }
 
         // POST: /Klient/Delete/5
@@ -162,10 +187,14 @@ namespace Silownia.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(long id)
         {
-            Klient klient = db.Klienci.Find(id);
-            db.Klienci.Remove(klient);
-            db.SaveChanges();
-            return RedirectToAction("Index", new { akcja = AkcjaEnum.UsunietoKlienta, info = klient.imieNazwisko });
+            if (Session["User"] != null)
+            {
+                Klient klient = db.Klienci.Find(id);
+                db.Klienci.Remove(klient);
+                db.SaveChanges();
+                return RedirectToAction("Index", new { akcja = AkcjaEnum.UsunietoKlienta, info = klient.imieNazwisko });
+            }
+            return HttpNotFound();
         }
 
         protected override void Dispose(bool disposing)
